style: torque + servo_torque test bodies to src/<module>/tests.rs per repo convention
